pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stewardme"
version = "0.1.0"
description = "Open-source AI steward that monitors the world and tells you what to do next"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
authors = [{name = "Raj Contractor"}]
keywords = ["ai", "rag", "personal-assistant", "self-hosted", "llm", "journal", "intelligence"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Scheduling",
]

dependencies = [
    "click>=8.1.0",
    "chromadb>=0.4.0",
    "anthropic>=0.18.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "pypdf>=4.2.0",
    "apscheduler>=3.10.0",
    "feedparser>=6.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "structlog>=23.2.0",
    "scikit-learn>=1.3.0",
    "pytrends>=4.9.2",
    "mcp[cli]>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/contractorr/stewardme"
Repository = "https://github.com/contractorr/stewardme"
Issues = "https://github.com/contractorr/stewardme/issues"

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
gemini = ["google-genai>=1.0.0"]
all-providers = ["openai>=1.0.0", "google-genai>=1.0.0"]
web = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.6",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
coach = "cli.main:cli"
stewardme-mcp = "coach_mcp.server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-q -ra --strict-markers"
markers = [
    "unit: fast isolated unit tests",
    "integration: integration tests spanning multiple components",
    "web: FastAPI route and app tests",
    "slow: tests intentionally excluded from fast default CI runs",
    "serial: tests that should not be parallelized",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src"]
omit = ["*/__main__.py"]

[tool.coverage.report]
fail_under = 55

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = true
